src/main/java/com/codehimalaye/springtaskcodehimalaye/v1/student/StudentController.java
package com.codehimalaye.springtaskcodehimalaye.v1.student;

import com.codehimalaye.springtaskcodehimalaye.v1.entity.Student;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/api/v1/students")
public class StudentController {
    @Autowired
    private StudentRepository studentRepository;

    @PostMapping
    public ResponseEntity<Student> createStudent(@Valid @RequestBody Student student) {
        try {
            Optional<Student> existingStudentWithRollNo = studentRepository.findByRollNo(student.getRollNo());

            if (existingStudentWithRollNo.isPresent()) {
                return new ResponseEntity<>(null, HttpStatus.CONFLICT);
            }

            Student savedStudent = studentRepository.save(student);

            return new ResponseEntity<>(savedStudent, HttpStatus.CREATED);
        } catch (Exception e) {
            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
        }
    }

    @GetMapping("")
    public ResponseEntity<List<Student>> getAllStudents() {
        List<Student> students = studentRepository.findAll();

        return new ResponseEntity<>(students, HttpStatus.OK);
    }

    @GetMapping("/{id}")
    public ResponseEntity<Student> getStudentById(@PathVariable("id") Long id) {
        try {
            Student student = studentRepository.findById(id).orElse(null);

            if (student == null) {
                return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
            }

            return new ResponseEntity<>(student, HttpStatus.OK);
        } catch (Exception e) {
            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
        }
    }

    @PutMapping("/{id}")
    public ResponseEntity<Student> updateStudent(@PathVariable("id") Long id, @RequestBody @Validated Student student) {
        Optional<Student> studentData = studentRepository.findById(id);

        if (studentData.isPresent()) {
            Student existingStudent = studentData.get();

            // Check if the rollNo is already in use by another student
            Optional<Student> existingStudentWithRollNo = studentRepository.findByRollNo(student.getRollNo());

            if (existingStudentWithRollNo.isPresent() && !existingStudentWithRollNo.get().getId().equals(existingStudent.getId())) {
                return new ResponseEntity<>(null, HttpStatus.CONFLICT);
            }

            // Update the existing student
            existingStudent.setName(student.getName());
            existingStudent.setAddress(student.getAddress());
            existingStudent.setRollNo(student.getRollNo());

            Student updatedStudent = studentRepository.save(existingStudent);

            return new ResponseEntity<>(updatedStudent, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
        }
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<HttpStatus> deleteStudent(@PathVariable("id") Long id) {
        try {
            studentRepository.deleteById(id);

            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
        } catch (Exception e) {
            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
        }
    }
}
